file rename from navigation-bar to navigationBar
